# Performance Optimization: Fast Column Setters