This PR changes the way that we do column lookups to be once-per-query rather than once-per row.  By doing this, we can speed up large query retrieval by ~500%.  

We have to do some funky stuff re: creating dynamic strongly-typed setter delegates, but other libraries have already gone this way (such as Protobuf), so we can follow in their footsteps a bit.

## Detailed explanation: Simple Syntax vs. Fast Speed

SQLite.net is very easy to use, and has great facility for mapping from SQLite tables into object properties.  However, for very large queries, response times can be a bit on the slow side:

	// *** Method 1 ***
	var orders = conn.Query<OrderDetail>("select * from [OrderDetail]");


If OrderDetail is 1 million records or so, then performance is very poor:

```Retrieved 1000000 records in 14852 ms```

Why does it take 15 seconds to query from a local database?  This should be FAST!

Traditionally this can be overcome by doing some low-level work and stepping through a SQLite query on your own:

	// *** Method 2 ***
	var stmt = SQLite3.Prepare2(conn.Handle, "select * from [OrderDetail]");
	var orders = new List<OrderDetail>();
	
	while (SQLite3.Step(stmt) == SQLite3.Result.Row)
	{
	    var id = SQLite3.ColumnInt(stmt, 0);
	    var orderId = SQLite3.ColumnInt(stmt, 1);
	    var productId = SQLite3.ColumnInt(stmt, 2);
	    var quantity = SQLite3.ColumnInt(stmt, 3);
	    var name = SQLite3.ColumnString(stmt, 4);
	
	    var orderDetail = new OrderDetail()
	    {
	        Id = id,
	        OrderId = orderId,
	        ProductId = productId,
	        Quantity = quantity,
	        Name = name
	    };
	    orders.Add(orderDetail);
	}

uuuugh.  It's much more cumbersome syntax, but the performance boost is undeniable:
```Retrieved 1000000 records in 1350 ms.```

That's more like it!  It runs in a tenth of the time!  Both are using SQLite.net with the same query on the same database -- so why is Method 1 so slow?

## In-Loop Reflection as the Root of Evil

The core of the problem lies in the fact that the inside of SQLite.net's `ExecuteDeferredQuery` loop makes a call to `PropertyInfo.SetValue ()`.  This is a rather heavy-weight call that uses Reflection every time it's called to make sure that the generic `object` being passed into it is compatible with the type held by that particular PropertyInfo.

So how do we fix this?  Is there a way to have the syntactic sugar of Method 1 with the speed of Method 2?

The good news is yes -- there are two alternatives that people have used to approach this problem.

### Fast-Member
The fastest method is to emit dynamic IL and link into that at runtime.  This is the method used by the [excellent Fast-Member library](https://github.com/mgravell/fast-member).

However, there are .NET platforms (such as Xamarin.iOS and Unity IL2CPP) that don't support such shenanigans, so I'd rather not go down that road.

### Strongly-Typed Delegates
The second method isn't quite as fast as dynamic IL, but it's still reasonably fast.  It involves doing all of the type-checking reflection _outside_ of the query loop, and creating strongly-typed delegates to set property values.  

Jon Skeet wrote [a blog post](https://codeblog.jonskeet.uk/2008/08/09/making-reflection-fly-and-exploring-delegates/) that explains this technique, as well as [a Stack Overflow answer that summarizes nicely](https://stackoverflow.com/a/1028058/12481680).  This is the same technique used to add [this same speed-boost to Google's Protobuf library](https://github.com/protocolbuffers/protobuf/blob/master/csharp/src/Google.Protobuf/Reflection/ReflectionUtil.cs).

I'm not doing everything quite the same way as he is (to be honest, I had a hard time following all of it), but I think the version I created is hopefully easy enough to read, and similarly zippy at runtime.

Note that we avoid quite a bit of confusion by simply skipping the Enum case, and falling back to the original method of simply calling Column.Set() on every row.  For enumerated types, then my pull request will simply fall back on the old (and slow) method of calling `PropertyInfo.SetValue()` on every row.  Yes, it's slow -- but at least it's not going to be any slower than it was before this change (and maybe someone else can help figure out the black-magic voodoo to make strongly-typed delegates for enums function).  

## Measuring Performance
So what does this all boil down to?  Well, let's check the performance.

Prior to my pull request:

	Method 1: Retrieved 1000000 records in 15009 ms.
	Method 2: Retrieved 1000000 records in 1419 ms.

After FastColumnSet:

	Method 1: Retrieved 1000000 records in 3087 ms.
	Method 2: Retrieved 1000000 records in 1426 ms.

Still not nearly as good as the hand-created mappings, but 15 seconds down to 3 seconds is still an impressive boost!

There's a lot of junk I needed to put in there to make Nullable types work properly (ugh).  If I skip Nullable checks and defer nullable types back to the legacy method, then it makes things a tad bit simpler, but I don't think the speed increase is that significant -- it shaves off maybe 500ms or so.

## Feedback?
What do you all think?  Is this clean enough / general-purpose enough to make it into the main trunk?  

I'm certainly open to input on how to make this all cleaner -- it was a bear getting this all to work properly, but I'm very very thankful for a comprehensive unit test suite in SQLite-net. Kudos for that. :)