@cherron-aptera 2019-12-18
Note: If anyone else would like to try reproducing my results, the program I used to time performance is here:

https://gist.github.com/cherron-aptera/a31f1de0fd489ffd74db248b62e5a6e5