@cherron-aptera 2020-08-12
Rebased and re-uploaded the branch to trigger the automatic build checks again.

This is a significant performance optimization, and I would still love to see this get merged into `master`.  I know there's a lot here and it's not super-easy to follow -- is there anything I can do to help to make the review process smoother, @praeclarum ?