@cherron-aptera 2021-05-06
@tranb3r It looks like it was attempting to persist a List<> with SqliteNetExtensions -- what type of object was stored in the list, do you know?

If you could possibly set a breakpoint and see what the expected method arguments are, I could look into why this is happening.

It looks like you're doing this on Android?  I'm curious to know if it happens on other platforms as well.  If we could build an isolated test project, that would help me with reproducing -- but I don't know how much trouble that would be.