@tranb3r 2021-05-06
A bit more context:
- yes this is on Android. I don't run this app on other platforms yet.
- there is no list involved. I'm getting recursively (with sqlitenetextensions) object A -> object B -> object C. Both relationships are ManyToOne.

I can definitely set a breakpoint, but could you please tell me where exactly (I guess when calling the CreateDelegate in CreateTypedSetterDelegate method ?) and what info you need ?

I can also try to build a repro project, but I guess it's only relevant to do it on android since the exception is in mono ?