@cherron-aptera 2021-05-06
Hmm, good question.  Maybe the first step is to see if we can reproduce it in a desktop environment?