@cherron-aptera 2021-05-06
@tranb3r If you are able to give me a code snippet for how you're using sqlitenetextensions, I can see about building a sample application from my end?