@tranb3r 2021-05-06
I've created a simple repro project, with 2 applications:
- TestSqliteNet.Android : xamarin forms android app
- TestSqliteNet.Net : netcoreapp3.1 console app

With sqlite-net-pcl 1.8.0-beta, there is an exception on both environments:
- android : `System.ArgumentException: method arguments are incompatible`
- netcoreapp3.1 : `System.ArgumentException: Cannot bind to the target method because its signature is not compatible with that of the delegate type`

No error with 1.7.335.

The code for the test is in TestClass.cs in the shared lib (TestSqliteNet).
It uses sqlite-net-pcl and SQLiteNetExtensions ; no need for sqlcipher nor async api.

Let me know if you need more information.
Thanks!

[TestSqliteNet.zip](https://github.com/praeclarum/sqlite-net/files/6437226/TestSqliteNet.zip)