@cherron-aptera 2021-05-10
I haven't yet taken time to try to fix this yet, but wanted to confirm that your test project is excellent -- I am able to run it locally and it fails for me with the same exception that you reported.  Well done on this -- thank you!