@tranb3r 2021-05-27
@cherron-aptera Did you have the time to take a look ? Maybe I should open a new issue ?