@tranb3r 2021-06-08
@cherron-aptera Sorry to insist: should I open a new issue ? Maybe somebody else will have the time to take a look. Thanks.